move animation of samples into separate Animation component (requires some changes due to mounting of components, e.g. check for samples undefined where used)
